Updated the Weights & Biases link since it was not a publicly accessible link
I had used the URL of the W&B report when I was editing it. It turns out that view is not accessible even if you make the report public, and you need to use the link generated with the "Share" button. I am updating this single link so that the report is accessible; nothing else has changed.